feat: workspace name flow and uniqueness checks
