+++ pkg/cmd/cmd.go
@@ -29,6 +29,7 @@ import (
 	. "github.com/daytonaio/daytona/pkg/cmd/targetconfig"
 	. "github.com/daytonaio/daytona/pkg/cmd/telemetry"
 	. "github.com/daytonaio/daytona/pkg/cmd/workspace"
+	. "github.com/daytonaio/daytona/pkg/cmd/workspace/create"
 	. "github.com/daytonaio/daytona/pkg/cmd/workspaceconfig"
 	"github.com/daytonaio/daytona/pkg/common"
 	"github.com/daytonaio/daytona/pkg/posthogservice"
