+++ pkg/cmd/workspace/code.go
@@ -11,13 +11,12 @@ import (
 	"strings"
 
 	"github.com/daytonaio/daytona/cmd/daytona/config"
-	"github.com/daytonaio/daytona/internal/jetbrains"
 	"github.com/daytonaio/daytona/internal/util"
 	apiclient_util "github.com/daytonaio/daytona/internal/util/apiclient"
 	"github.com/daytonaio/daytona/pkg/apiclient"
-	"github.com/daytonaio/daytona/pkg/ide"
+	workspace_common "github.com/daytonaio/daytona/pkg/cmd/workspace/common"
+	"github.com/daytonaio/daytona/pkg/cmd/workspace/create"
 	"github.com/daytonaio/daytona/pkg/server/workspaces"
-	"github.com/daytonaio/daytona/pkg/telemetry"
 	ide_views "github.com/daytonaio/daytona/pkg/views/ide"
 	"github.com/daytonaio/daytona/pkg/views/workspace/selection"
 
@@ -78,8 +77,8 @@ var CodeCmd = &cobra.Command{
 				return err
 			}
 		}
-		if ideFlag != "" {
-			ideId = ideFlag
+		if create.IdeFlag != "" {
+			ideId = create.IdeFlag
 		}
 
 		if ws.State != nil || ws.State.Uptime < 1 {
@@ -94,7 +93,7 @@ var CodeCmd = &cobra.Command{
 
 		providerMetadata := *ws.Info.ProviderMetadata
 
-		gpgKey, err := GetGitProviderGpgKey(apiClient, ctx, providerConfigId)
+		gpgKey, err := workspace_common.GetGitProviderGpgKey(apiClient, ctx, providerConfigId)
 		if err != nil {
 			log.Warn(err)
 		}
@@ -102,40 +101,12 @@ var CodeCmd = &cobra.Command{
 		yesFlag, _ := cmd.Flags().GetBool("yes")
 		ideList := config.GetIdeList()
 		ide_views.RenderIdeOpeningMessage(ws.TargetId, ws.Name, ideId, ideList)
-		return openIDE(ideId, activeProfile, ws.Id, providerMetadata, yesFlag, gpgKey)
+		return workspace_common.OpenIDE(ideId, activeProfile, ws.Id, providerMetadata, yesFlag, gpgKey)
 	},
 	ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 		return getWorkspaceNameCompletions()
 	},
 }
-
-func openIDE(ideId string, activeProfile config.Profile, workspaceId string, workspaceProviderMetadata string, yesFlag bool, gpgKey string) error {
-	telemetry.AdditionalData["ide"] = ideId
-
-	switch ideId {
-	case "vscode":
-		return ide.OpenVSCode(activeProfile, workspaceId, workspaceProviderMetadata, gpgKey)
-	case "ssh":
-		return ide.OpenTerminalSsh(activeProfile, workspaceId, gpgKey, nil)
-	case "browser":
-		return ide.OpenBrowserIDE(activeProfile, workspaceId, workspaceProviderMetadata, gpgKey)
-	case "cursor":
-		return ide.OpenCursor(activeProfile, workspaceId, workspaceProviderMetadata, gpgKey)
-	case "jupyter":
-		return ide.OpenJupyterIDE(activeProfile, workspaceId, workspaceProviderMetadata, yesFlag, gpgKey)
-	case "fleet":
-		return ide.OpenFleet(activeProfile, workspaceId, gpgKey)
-	default:
-		_, ok := jetbrains.GetIdes()[jetbrains.Id(ideId)]
-		if ok {
-			return ide.OpenJetbrainsIDE(activeProfile, ideId, workspaceId, gpgKey)
-		}
-	}
-
-	return errors.New("invalid IDE. Please choose one by running `daytona ide`")
-}
-
-var ideFlag string
 
 func init() {
 	ideList := config.GetIdeList()
@@ -144,7 +115,7 @@ func init() {
 		ids[i] = ide.Id
 	}
 	ideListStr := strings.Join(ids, ", ")
-	CodeCmd.Flags().StringVarP(&ideFlag, "ide", "i", "", fmt.Sprintf("Specify the IDE (%s)", ideListStr))
+	CodeCmd.Flags().StringVarP(&create.IdeFlag, "ide", "i", "", fmt.Sprintf("Specify the IDE (%s)", ideListStr))
 
 	CodeCmd.Flags().BoolVarP(&yesFlag, "yes", "y", false, "Automatically confirm any prompts")
