+++ pkg/cmd/workspace/common/configuration_flags.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Daytona Platforms Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package util
+package common
 
 import (
 	"fmt"
