+++ pkg/cmd/workspace/create/add_from_config.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Daytona Platforms Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package util
+package create
 
 import (
 	"context"
