+++ pkg/cmd/workspace/ssh.go
@@ -10,6 +10,7 @@ import (
 	"github.com/daytonaio/daytona/internal/util"
 	apiclient_util "github.com/daytonaio/daytona/internal/util/apiclient"
 	"github.com/daytonaio/daytona/pkg/apiclient"
+	workspace_common "github.com/daytonaio/daytona/pkg/cmd/workspace/common"
 	"github.com/daytonaio/daytona/pkg/ide"
 	views_util "github.com/daytonaio/daytona/pkg/views/util"
 	"github.com/daytonaio/daytona/pkg/views/workspace/selection"
@@ -81,7 +82,7 @@ var SshCmd = &cobra.Command{
 			sshArgs = append(sshArgs, args[1:]...)
 		}
 
-		gpgKey, err := GetGitProviderGpgKey(apiClient, ctx, providerConfigId)
+		gpgKey, err := workspace_common.GetGitProviderGpgKey(apiClient, ctx, providerConfigId)
 		if err != nil {
 			log.Warn(err)
 		}
